Move Name to engine/types.rs
It will be moving out of StratEngine-only use to general use in the Engine
traits.
