+++ src/engine/mod.rs
@@ -21,6 +21,7 @@ pub use self::types::FilesystemUuid;
 pub use self::types::PoolUuid;
 pub use self::types::Redundancy;
 pub use self::types::RenameAction;
+pub use self::types::Name;
 
 #[macro_use]
 mod macros;
