+++ src/engine/structures.rs
@@ -2,49 +2,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::borrow::Borrow;
 use std::collections::{hash_map, HashMap};
-use std::fmt;
 use std::iter::IntoIterator;
-use std::ops::Deref;
-use std::rc::Rc;
 
 use uuid::Uuid;
 
-use super::engine::{HasName, HasUuid};
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone)]
-pub struct Name(Rc<String>);
-
-impl Name {
-    pub fn new(name: String) -> Name {
-        Name(Rc::new(name))
-    }
-
-    pub fn to_owned(&self) -> String {
-        self.0.deref().to_owned()
-    }
-}
-
-impl Deref for Name {
-    type Target = str;
-
-    fn deref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl Borrow<str> for Name {
-    fn borrow(&self) -> &str {
-        &**self.0
-    }
-}
-
-impl fmt::Display for Name {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
+use engine::Name;
 
 /// Map UUID and name to T items.
 #[derive(Debug)]
